Handle LookupError properly
Address #109 and #117 : for these, in expand_EXPN, nltk requires 'averaged_perceptron_tagger' and 'universal_tagset' to be downloaded, but expand_EXPN has a bare except that just returns the original nsw if anything fails.  This obscured the missing data dependencies and meant that expand_EXPN never did anything. By explicitly raising LookupError , we can surface this issue to the user. It would be better to handle this dependency for the user, but this at least makes the issue transparent